{{cookiecutter.project}}/requirements.txt
Django==2.2.10
django-model-utils==3.2.0
django-redis==4.10.0  # https://github.com/niwinz/django-redis
djangorestframework==3.9.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.0.2
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
django-celery-results==1.1.2
django-rest-swagger==2.2.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0

pytz==2019.1
PyYAML==5.1.1
sqlparse==0.3.0
redis==3.2.1  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
flower==0.9.3  # https://github.com/mher/flower
mysqlclient==1.4.2.post1
pycrypto==2.6.1
ConcurrentLogHandler==0.9.1
gunicorn==19.9.0
kombu==4.6.3
coreapi==2.3.3
Pillow==6.2.1
openpyxl==3.0.2

